+++ src/Common/utils/types.ts
@@ -6,10 +6,10 @@ export interface TestableComponentProps {
     dataTestId? : string;
 }
 
-export function getTestableComponentProps( props : TestableComponentProps ): { "data-testId"? : string} {
+export function getTestableComponentProps( props : TestableComponentProps ): { "data-testid"? : string} {
     var testableComponentProps = {} ;
     if(props.dataTestId) {
-        testableComponentProps["data-testId"] = props.dataTestId ;
+        testableComponentProps["data-testid"] = props.dataTestId ;
     }
     return testableComponentProps ;
 }